Project: ETCH-A-SKETCH
I had to create the etch-a-sketch project from The Odin Project. The project took me about a day and a half of work. I solidified a lot of the concepts related to DOM manipulation as well as arrays and loops. This was definitely a challenge. But I got through it.